tests/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
redis==5.0.1
pytest==7.4.2
pytest-asyncio==0.21.1
